app/requirements.txt
fastapi
uvicorn
openai
ollama
python-dotenv
requests
sqlite3
google-generativeai